Update Bos taurus Angus example
- Two estimation settings: fast and end-to-end
- Auto setting for MATLAB calls
