+++ examples/Bos_taurus_Angus/multitier/entity_vs_tier.csv
@@ -2,23 +2,23 @@ tier,entity,breed,diet,individual
 breed,male,male,,
 diet,CTRL,male,CTRL,
 diet,TMR,male,TMR,
-individual,PT524956505,male,CTRL,PT524956505
+individual,PT933843912,male,TMR,PT933843912
 individual,PT333653651,male,TMR,PT333653651
-individual,PT533843896,male,CTRL,PT533843896
-individual,PT833653644,male,TMR,PT833653644
-individual,PT724523831,male,TMR,PT724523831
-individual,PT333842562,male,CTRL,PT333842562
-individual,PT533358890,male,TMR,PT533358890
-individual,PT924401183,male,CTRL,PT924401183
-individual,PT233843883,male,TMR,PT233843883
-individual,PT433843806,male,TMR,PT433843806
-individual,PT933843894,male,CTRL,PT933843894
-individual,PT933602927,male,CTRL,PT933602927
 individual,PT524401180,male,CTRL,PT524401180
 individual,PT224401177,male,CTRL,PT224401177
-individual,PT624139868,male,TMR,PT624139868
-individual,PT033634130,male,TMR,PT033634130
-individual,PT833653649,male,CTRL,PT833653649
+individual,PT924401183,male,CTRL,PT924401183
+individual,PT333842562,male,CTRL,PT333842562
+individual,PT533358890,male,TMR,PT533358890
+individual,PT524956505,male,CTRL,PT524956505
 individual,PT533987885,male,CTRL,PT533987885
-individual,PT933843912,male,TMR,PT933843912
+individual,PT233843883,male,TMR,PT233843883
+individual,PT833653649,male,CTRL,PT833653649
+individual,PT533843896,male,CTRL,PT533843896
+individual,PT624139868,male,TMR,PT624139868
+individual,PT724523831,male,TMR,PT724523831
+individual,PT433843806,male,TMR,PT433843806
+individual,PT833653644,male,TMR,PT833653644
+individual,PT933843894,male,CTRL,PT933843894
 individual,PT424401157,male,TMR,PT424401157
+individual,PT933602927,male,CTRL,PT933602927
+individual,PT033634130,male,TMR,PT033634130
